Few small changes
Mainly, wrapping was off because of hard `\n`'s in the file.  E.g., 
```We will make use of the NestJS CLI to get started quickly. If you have not 
installed it, use
the following command:```

So I removed those.  Also made a couple small grammar and typo fixes.

Great stuff!!!!